requirements.txt
appdirs==1.4.4
asn1crypto==0.24.0
certifi==2018.8.24
chardet==3.0.4
cryptography==2.6.1
devscripts==2.19.5+deb10u1
distro==1.3.0
distro-info==0.21
dnspython==1.16.0
entrypoints==0.3
fs==2.4.13
gpg==1.12.0
httplib2==0.11.3
idna==2.6
iotop==0.6
keyring==17.1.1
keyrings.alt==3.1.1
meld==3.20.0
pexpect==4.6.0
pycairo==1.16.2
pycrypto==2.6.1
pycurl==7.43.0.2
PyGObject==3.30.4
PySimpleSOAP==1.16.2
python-apt==1.8.4.3
python-debian==0.1.35
python-debianbts==2.8.2
python-distutils-extra==2.39
python-gnupg==0.4.7
python-magic==0.4.16
pytz==2021.1
pyxdg==0.25
PyYAML==5.4.1
reportbug===7.5.3-deb10u1
requests==2.21.0
SecretStorage==2.3.1
six==1.12.0
ufw==0.36
unattended-upgrades==0.1
unidiff==0.5.4
urllib3==1.24.1
